clippy

+++ src/health.rs
@@ -1261,8 +1261,10 @@ fn disk_space(path: &str) -> (u64, u64) {
         unsafe {
             let mut stat: libc::statvfs = std::mem::zeroed();
             if libc::statvfs(c_path.as_ptr(), &mut stat) == 0 {
-                let total = stat.f_blocks as u64 * stat.f_frsize;
-                let free = stat.f_bavail as u64 * stat.f_frsize;
+                #[allow(clippy::unnecessary_cast)]
+                let total = stat.f_blocks as u64 * stat.f_frsize as u64;
+                #[allow(clippy::unnecessary_cast)]
+                let free = stat.f_bavail as u64 * stat.f_frsize as u64;
                 return (total, free);
             }
         }
